requirements.txt
pandas==2.1.1
numpy==1.26.4
openpyxl==3.1.2
SQLAlchemy==2.0.21
PyYAML==6.0.1
loguru==0.7.2
pytest==7.4.3